feat: Handle materialized view through alembic
The FilteredAggreation view used to be created manually, making it
out of the alembic migration scope.
Now, the view is automatically handled in alembic and can be safely
migrated just like any other table.
Note a materialized view might require to recreate from scratch, which
can take some time. A workaround is detailed in the README to prevent
any service interruption.
