+++ requirements.txt
@@ -1,4 +1,5 @@
 alembic==1.6.5
+alembic-utils==0.7.6
 attrs==21.2.0
 autopep8==1.5.7
 blinker==1.4
@@ -11,6 +12,7 @@ Flask-Caching==1.10.1
 Flask-HTTPAuth==4.4.0
 Flask-Migrate==3.0.1
 Flask-SQLAlchemy==2.5.1
+flupy==1.1.9
 greenlet==1.1.0
 gunicorn==20.1.0
 idna==2.10
@@ -23,6 +25,7 @@ mccabe==0.6.1
 numpy==1.20.3
 packaging==20.9
 pandas==1.2.4
+parse==1.19.0
 pluggy==0.13.1
 psycopg2-binary==2.8.6
 py==1.10.0
@@ -43,6 +46,7 @@ SQLAlchemy==1.4.15
 tabulate==0.8.9
 text-unidecode==1.3
 toml==0.10.2
+typing-extensions==4.0.1
 urllib3==1.26.5
 Werkzeug==2.0.1
 wrapt==1.12.1
